Added local storage check for existing nfts, added session storage, added methodology for determining price paid

+++ DataHandlers/LocalStorage.js
@@ -1,24 +1,26 @@
+const entryExists = (currentState, newState) => {
+  return currentState.filter((entry) => entry.txId === newState.txId)
+    ? true
+    : false;
+};
+
 export const localStorageHandler = (walletAddr, payload) => {
+  sessionStorage.setItem("user", walletAddr); //Persist wallet adress between page refreshes
   if (localStorage.getItem(walletAddr) === null) {
-    localStorage.setItem(walletAddr, Object.assign(JSON.stringify([payload])));
+    localStorage.setItem(walletAddr, Object.assign(JSON.stringify([payload]))); //inititialize walletAddr to be an array with payload object at first index
     return;
   }
   const currentState = JSON.parse(localStorage.getItem(walletAddr));
 
-  const newState = Object.assign(currentState, currentState.push(payload));
-  console.log(newState);
+  if (entryExists(currentState, payload)) {
+    console.log("Entry is already in localStorage");
+    // Location for setting error message
+    return;
+  }
 
-  const newStateStr = JSON.stringify(newState);
+  const newStateStr = JSON.stringify(
+    Object.assign(currentState, currentState.push(payload))
+  );
 
   localStorage.setItem(walletAddr, newStateStr);
-
-  //   Object.assign();
-};
-
-const sampleDataSet = {
-  walletAddr: "wallet adress",
-  nfts: [
-    { txId: "txId", name: "name", src: "src" },
-    { txId: "txId", name: "name", src: "src" },
-  ],
 };
